+++ example/meteor-server/start.sh
@@ -1,0 +1,1 @@
+METEOR_PACKAGE_DIR=../ && meteor