+++ meteor-apollo-accounts/src/getLoginToken.js
@@ -1,3 +1,0 @@
-import {getLoginToken} from './store'
-
-export default getLoginToken