+++ meteor-apollo-accounts/src/store.js
@@ -1,6 +1,19 @@
+const onChangeCallbacks = []
+
 const tokenDidChange = function () {
   // Looking for a better way to refetch all queries
   // window.location.reload(0)
+  for (const callback of onChangeCallbacks) {
+    try {
+      callback()
+    } catch (error) {
+      console.log('Error on onChangeCallback', error)
+    }
+  }
+}
+
+export const onTokenChange = function (callback) {
+  onChangeCallbacks.push(callback)
 }
 
 export const storeLoginToken = function (userId, token, tokenExpires) {