+++ meteor-apollo-accounts/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+### v1.1.0
+
+- onTokenChange callbacks
+- Login with facebook
+- Login with google
+
+
 ### v1.0.6
 
 - Disable reload on login or logout